Cleanup post endpoint
