+++ laa_court_data_api_app/routers/laa_references.py
@@ -8,7 +8,7 @@ from laa_court_data_api_app.internal.court_data_adaptor_client import CourtDataA
 from laa_court_data_api_app.models.laa_references.external.request.laa_references_patch_request import \
     LaaReferencesPatchRequest as ExternalPatchRequest
 from laa_court_data_api_app.models.laa_references.external.request.laa_references_post_request import \
-    LaaReferencesPost as ExternalPostRequest
+    LaaReferencesPostRequest as ExternalPostRequest
 from laa_court_data_api_app.models.laa_references.external.response.laa_references_error_response import \
     LaaReferencesErrorResponse
 from laa_court_data_api_app.models.laa_references.internal.request.laa_references_patch_request import \
