+++ test/routers/laa_references/post_test.py
@@ -5,6 +5,8 @@ from fastapi.testclient import TestClient
 
 from laa_court_data_api_app.internal.oauth_client import OauthClient
 from laa_court_data_api_app.main import app
+from laa_court_data_api_app.models.laa_references.external.request.laa_references_post_request import \
+    LaaReferencesPostRequest
 
 from ..fixtures import *
 
@@ -21,7 +23,7 @@ def test_laa_references_post_returns_accepted(mock_settings, mock_cda_settings, 
     mock_cda_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="pass-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="pass-u", maat_reference=1234567).json())
 
     assert response.status_code == 202
     assert response.content == b''
@@ -38,7 +40,7 @@ def test_laa_references_post_returns_bad_request(mock_settings, mock_cda_setting
     mock_cda_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="fail-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="fail-u", maat_reference=1234567).json())
 
     assert response.status_code == 400
     assert mock_cda_client["laa_references_post_fail_route"].called
@@ -55,7 +57,7 @@ def test_laa_references_post_returns_not_found(mock_settings, mock_cda_settings,
     mock_cda_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="notfound-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="notfound-u", maat_reference=1234567).json())
 
     assert response.status_code == 404
     assert response.content == b''
@@ -72,7 +74,7 @@ def test_laa_references_post_returns_unprocessable_entity(mock_settings, mock_cd
     mock_cda_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="unprocessable-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="unprocessable-u", maat_reference=1234567).json())
 
     assert response.status_code == 422
     assert mock_cda_client["laa_references_post_unprocessable_entity_route"].called
@@ -89,7 +91,7 @@ def test_laa_references_post_returns_server_error(mock_settings, mock_cda_settin
     mock_cda_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="servererror-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="servererror-u", maat_reference=1234567).json())
 
     assert response.status_code == 424
     assert response.content == b''
@@ -107,7 +109,7 @@ def test_laa_references_post_returns_none(mock_settings, mock_cda_settings, over
     mock_settings.return_value = override_get_cda_settings
 
     response = client.post("/v2/laa_references",
-                           data=LaaReferencesPost(user_name="test-u", maat_reference=1234567).json())
+                           data=LaaReferencesPostRequest(user_name="test-u", maat_reference=1234567).json())
 
     assert response.status_code == 424
     assert response.content == b''
